Make cppcheck happy (#21)
Ran cppcheck on the kernel code and fixed some issues. Some remain:

* Unused functions, some because of partially implemented features,
others because cppcheck cannot figure out that they are called from
assembly language functions.
*
[Here](https://github.com/phaubertin/jinue/blob/a6fef92282482cb0227392fd854a8c27396622f2/kernel/hal/vm_pae.c#L451),
cppcheck reports that `addr_space` may be `NULL`. There may be a way to
clean up that part of the code. Strangely, though, this warning
disappears when we pass `-D___GNUC___=4` or `-D__CPPCHECK__=1` to
cppcheck, even though these macros are used nowhere in the code. (This
was found when attempting a solution for the next issue.)
*
[Here](https://github.com/phaubertin/jinue/blob/a6fef92282482cb0227392fd854a8c27396622f2/kernel/hal/vm.c#L362-L366),
cppcheck does not know that panic() never returns, so it flags a
potential dereference of a `NULL` `phdr`. What I tried:

* Specify `panic()` does not return in a cppcheck library file. I was
not able to make this work.
* Define a `NORETURN` macro that is conditionally `__attribute__
((noreturn))` when `__GNUC__` or `__CPPCHECK__` is set, and add this
attribute to the `panic()` declaration. This worked, but also made an
unrelated warning disappear as mentioned above. I was not comfortable
with this solution because I don't know what relevant warnings might
disappear in the future.
* Set `__attribute__ ((noreturn))` directly on the `panic()`
declaration. This works but would make the code less portable across
compilers.

More work will be needed in order to fix the remaining issues.