Rename JOIN_THREAD to AWAIT_THREAD (#81)
Rename `JOIN_THREAD` system call to `AWAIT_THREAD`: the system call
performs the awaiting part of `pthread_join()` but does not deal with
the exit status so it's more of a call upon which `pthread_join()` is
built rather than a system call thinly wrapped by this function, so
there is no real need to follow the same naming. Await is more
descriptive of what it does than "join".