Fix kernel pointer offset used by SYSCALL instruction (#90)
This PR fixes an issue introduced by #83 where the trap code for system
calls that use the `SYSCALL` instruction do not use the correct offset
in the per-CPU data structure when it retrieves the kernel stack
pointer. This was caused by a change to the `percpu_t` structure that
wasn't reflected in the offset the assembly language code uses.