Replace KLIMIT usage with KERNEL_BASE and JINUE_KLIMIT (#123)
* Replace all usage of constant `KLIMIT` with either of the following,
as appropriate:
  * `JINUE_KLIMIT`, the limit between user space and kernel space.
* `KERNEL_BASE`, the address where the kernel image is mapped in kernel
space
* Delete constant `KLIMIT`.

This is preparation work for moving the kernel mapping area. Right now,
the kernel image is mapped at the very beginning of kernel space, which
means `KERNEL_BASE` equals `JINUE_KLIMIT`. However, the mapping area
will be moved to before the kernel image, so this will no longer be the
case, hence the need for two distinct constants.