Set up final kernel page tables from setup code (#127)
* Have the 32-bit setup code determine if Physical Address Extension
(PAE) is supported and enable it if it is.
* Have the 32-bit setup code set up the final page tables for the kernel
address space.
* Do not move the kernel image from 0x100000 (1MB) to 0x1000000 (16MB)
anymore, but perform all memory allocations starting at 0x1000000
(16MB).

The current situation is this: the setup code sets up temporary page
tables for the kernel early boot which map only a part of the kernel
address space as well as the first two megabytes of memory and a few
megabytes starting at 0x1000000 (16MB). Once the kernel gets control, it
modifies or replaces them with equivalent ones when it moves the kernel
image from 0x100000 (1MB) to 0x1000000 (16MB) and when it enables PAE.
When it does that, it also moves the initial memory allocations which
are right after the kernel image. Finally, it completely discards these
temporary page tables and creates the final page tables for the kernel
address space.

The rationale for this situation was as follow:

* Do as little work as possible in the setup code since its size is
limited.
* Do as much work as possible in the kernel proper where all the kernel
code is available, including some functions from the standard C library
and other utility functions.
* Do as little work as possible before logging (VGA, UART) is enabled
for observability.
* Regarding moving the kernel image in memory: use as little memory as
possible under the 16MB mark to leave it available for DMA.

However, this leads to more overall complexity where making any change
to the kernel address space layout becomes tedious, since multiple page
table building steps must be adapted. Furthermore, it introduces more
dependencies between steps of the boot process.