Reassign X86_PTE_PROT_NONE to bit 10 (#138)
Reassign the `X86_PTE_PROT_NONE` flag from bit 11 to bit 10 in page
tables entry. This flag is used to represent a page that is present but
unreadable. Bit 11 is assigned by the architecture for use with HLAT
paging. This is not something we support, so bit 11 is actually fine,
but using a bit that is never used for any purpose is more future proof.