Kernel log ring buffer (#141)
Write logs into a ring buffer before output to serial port and/or VGA.

The immediate advantage of doing this is that we can log messages during
early boot initialization, before the serial port and VGA are
initialized. Once they are, they will output the logs from the buffer.
This reduces dependencies between steps during initialization.

In the longer term, this will permit the implementation of a system call
to get the logs from user space and, possibly, to decouple log producers
from log consumers by making logging asynchronous.